tools/rp_storage_tool: remove unused partial adl impl
This was mostly unused and the serialization half was
incomplete: retain the deserialization part in a utility
file for where it's used to read RecordBatchHeader.
