--- tools/rp_storage_tool/redpanda-adl/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "redpanda-adl"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-serde = "1.0"
-log = "0.4.17"
-bincode = "2.0.0-rc.2"
