--- tools/rp_storage_tool/redpanda-adl/src/lib.rs
@@ -1,7 +1,0 @@
-mod de;
-mod error;
-//mod se;
-
-pub use de::{from_bytes, Deserializer};
-pub use error::{Error, Result};
-//pub use se::to_bytes;
