--- tools/rp_storage_tool/redpanda-adl/src/se.rs
@@ -1,197 +1,0 @@
-use crate::error::{EncodeError, Result};
-use serde::ser;
-use serde::Serialize;
-
-struct Serializer<'a> {
-    output: &'a mut Box<dyn std::io::Write>,
-}
-
-pub fn to_bytes<T>(value: &T) -> Result<Vec<u8>>
-where
-    T: Serialize,
-{
-    let mut out_vec = Box::new(Vec::<u8>::new());
-    let mut serializer = Serializer {
-        output: &mut out_vec,
-    };
-    value.serialize(&mut serializer)?;
-    drop(serializer);
-    Ok(out_vec.into())
-}
-
-impl<'a, 'b> ser::Serializer for &'a mut Serializer<'b> {
-    type Ok = ();
-    type Error = EncodeError;
-    type SerializeSeq = Self;
-    type SerializeTuple = Self;
-    type SerializeTupleStruct = Self;
-    type SerializeTupleVariant = Self;
-    type SerializeMap = Self;
-    type SerializeStruct = Self;
-    type SerializeStructVariant = Self;
-
-    fn serialize_bool(self, v: bool) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_i8(self, v: i8) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_i16(self, v: i16) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_i32(self, v: i32) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_i64(self, v: i64) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_u8(self, v: u8) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_u16(self, v: u16) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_u32(self, v: u32) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_u64(self, v: u64) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_f32(self, v: f32) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_f64(self, v: f64) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_char(self, v: char) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_str(self, v: &str) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_bytes(self, v: &[u8]) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_none(self) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_some<T: ?Sized>(self, value: &T) -> std::result::Result<Self::Ok, Self::Error>
-    where
-        T: Serialize,
-    {
-        todo!()
-    }
-
-    fn serialize_unit(self) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_unit_struct(
-        self,
-        name: &'static str,
-    ) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_unit_variant(
-        self,
-        name: &'static str,
-        variant_index: u32,
-        variant: &'static str,
-    ) -> std::result::Result<Self::Ok, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_newtype_struct<T: ?Sized>(
-        self,
-        name: &'static str,
-        value: &T,
-    ) -> std::result::Result<Self::Ok, Self::Error>
-    where
-        T: Serialize,
-    {
-        todo!()
-    }
-
-    fn serialize_newtype_variant<T: ?Sized>(
-        self,
-        name: &'static str,
-        variant_index: u32,
-        variant: &'static str,
-        value: &T,
-    ) -> std::result::Result<Self::Ok, Self::Error>
-    where
-        T: Serialize,
-    {
-        todo!()
-    }
-
-    fn serialize_seq(
-        self,
-        len: Option<usize>,
-    ) -> std::result::Result<Self::SerializeSeq, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_tuple(self, len: usize) -> std::result::Result<Self::SerializeTuple, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_tuple_struct(
-        self,
-        name: &'static str,
-        len: usize,
-    ) -> std::result::Result<Self::SerializeTupleStruct, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_tuple_variant(
-        self,
-        name: &'static str,
-        variant_index: u32,
-        variant: &'static str,
-        len: usize,
-    ) -> std::result::Result<Self::SerializeTupleVariant, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_map(
-        self,
-        len: Option<usize>,
-    ) -> std::result::Result<Self::SerializeMap, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_struct(
-        self,
-        name: &'static str,
-        len: usize,
-    ) -> std::result::Result<Self::SerializeStruct, Self::Error> {
-        todo!()
-    }
-
-    fn serialize_struct_variant(
-        self,
-        name: &'static str,
-        variant_index: u32,
-        variant: &'static str,
-        len: usize,
-    ) -> std::result::Result<Self::SerializeStructVariant, Self::Error> {
-        todo!()
-    }
-}
